Fixed a breaking bug
In `common.hpp` the `load_be` and `load_le` functions were broken. Instead
of reading in bytes in their respective order, these functions would output
the last byte it was supposed to read with the rest set to 0xff's. I fixed
this by explicitly casting the value read from the buffer to uint8_t, since
that's what the buffer (should) contain. This fixed the issue. I am not
quite sure why this happened though..

Signed-off-by: cytesys <35925878+cytesys@users.noreply.github.com>