Refactor hexdigest() to return std::string
The old hexdigest() "returned" a c-string, which was unconvenient.
In addition to that, hexdigest() had almost the exact same code as
digest(), but had an additional step (snprintf). Code reuse is bad, and
it's already an issue in this project, because every hash class here
shares a lot of code. This is something that I'm already working on.

In any case, hexdigest() now returns a proper std::string. It's
inefficient/slow, I'm sure, but hexdigest() should only be used when
printing to the terminal or when storing in a text file anyway.

Signed-off-by: cytesys <35925878+cytesys@users.noreply.github.com>