Trying to actually use the newly created HBase class
This HBase class should be able to provide common functionality for every
hash class.

Templated classes are a pain to export. The way it's exported right now is
probably not good (or even correct?), but it works and doesn't raise any
errors or warnings from my local IDE and compiler.

Signed-off-by: cytesys <35925878+cytesys@users.noreply.github.com>