Updated the Blake hash classes to use HBase (plus more)
The Blake hash classes now inherits from an intermediate _BlakeImpl class,
which in turn inherits from HBase. Made unit tests for all the final Blake
classes as well. In addition, a bug with the original Blake implementation
were fixed. If the last block only consists of padding, the message length
should not be xored together with the last four constants in the working
array.

A lot of setters and getters were added to HBase, and every HBase derived
classes were updated accordingly.
All the unit tests for Hash classes were refactored a bit for better
readability and to perform more extensive testing. Unit tests for the ROT
macros were also added.

Signed-off-by: cytesys <35925878+cytesys@users.noreply.github.com>