Updated BLAKE2 to use HBase + unit tests for it.
This BLAKE2 implementation does have a bug though; when the last input
block is a full block, it get processed right away. BLAKE2 however, needs
to know if this was the final block, and my implementation has no way of
knowing that as it is now. I suspect the last block would have to be
buffered, and then processed on the NEXT call to either update() or
finalize(), but I have a feeling that this might break a lot of other
stuff... We'll see :]

Signed-off-by: cytesys <35925878+cytesys@users.noreply.github.com>